chore: [OFFNAL-150] [KanuKim97] : 코드 리뷰 반영
